+++ config/config.php
@@ -6,7 +6,15 @@ return [
 
     // Obscure IDs in certain access areas
     'obscure' => [
-        'adminarea',
+        'rotate' => false,
+        'areas' => [
+            'adminarea',
+        ],
+    ],
+
+    // Allow search engines to index access areas
+    'indexable' => [
+        'frontarea',
     ],
 
     // Global Route Override