+++ resources/views/adminarea/pages/datatable-index.blade.php
@@ -23,6 +23,7 @@
 
                     <div class="box">
                         <div class="box-body">
+                            @yield('datatable-filters')
                             {!! $dataTable->table(['class' => 'table table-striped table-hover responsive dataTableBuilder', 'id' => "{$id}"]) !!}
                         </div>
                     </div>