+++ resources/views/common/errors/401.blade.php
@@ -1,0 +1,5 @@
+@extends('cortex/foundation::common.errors.layout')
+
+@section('title', __('Unauthorized'))
+@section('code', '401')
+@section('message', __('Unauthorized'))