+++ CHANGELOG.md
@@ -5,6 +5,21 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](CONTRIBUTING.md).
 
 
+## [v3.1.1] - 2019-12-04
+- Add ajax filters capabilities to datatables
+- Tweak obscure feature
+- Add id attribute to the main div content in body
+- Fix file size validation rule
+
+## [v3.1.0] - 2019-11-23
+- Fix Crawling Robots indexable access areas feature
+- Rename AccessArea Middleware to SetAccessArea
+- Arrange SetAccessArea & Reauthenticate middleware priorities order
+- Update error views
+- Override ThrottleRequests middleware
+- Handle ThrottleRequestsException
+- Remember current URL for later redirect, when unauthenticated exception handled
+
 ## [v3.0.3] - 2019-10-14
 - Update menus & breadcrumbs event listener to accessarea.ready
 - Fix wrong dependencies letter case
@@ -118,6 +133,8 @@ This project adheres to [Semantic Versioning](CONTRIBUTING.md).
 ## v0.0.1 - 2017-03-14
 - Tag first release
 
+[v3.1.1]: https://github.com/rinvex/cortex-foundation/compare/v3.1.0...v3.1.1
+[v3.1.0]: https://github.com/rinvex/cortex-foundation/compare/v3.0.3...v3.1.0
 [v3.0.3]: https://github.com/rinvex/cortex-foundation/compare/v3.0.2...v3.0.3
 [v3.0.2]: https://github.com/rinvex/cortex-foundation/compare/v3.0.1...v3.0.2
 [v3.0.1]: https://github.com/rinvex/cortex-foundation/compare/v3.0.0...v3.0.1